Struct pyscemu::Emu

pub struct Emu { /* private fields */ }

Implementations

impl Emu

pub fn write_wide_string(&mut self, to: u64, from: &str)

write a wide string on memory.

pub fn write_buffer(&mut self, to: u64, from: &[u8])

write a python list of bytes to the emulator memory.

pub fn read_buffer(&mut self, from: u64, sz: usize) -> PyResult<Vec<u8>>

read a buffer from the emulator memory to a python list of bytes.

pub fn print_maps_by_keyword(&self, kw: &str)

print all the maps that match a substring of the keyword provided.

pub fn print_maps(&self)

print all the memory maps on the process address space.

pub fn get_addr_base(&self, addr: u64) -> PyResult<u64>

get the base address of a given address. Will make an exception if it’s invalid address.

pub fn is_mapped(&self, addr: u64) -> PyResult<bool>

this method checks if the given address is allocated or not.

pub fn get_addr_name(&self, addr: u64) -> PyResult<String>

get the memory map name where is the given address. Will cause an exception if the address is not allocated.

pub fn dump(&self, addr: u64)

visualize the bytes on the given address.

pub fn dump_n(&self, addr: u64, amount: u64)

visualize the amount of bytes provided on address.

pub fn dump_qwords(&self, addr: u64, n: u64)

visualize a number of qwords on given address.

pub fn dump_dwords(&self, addr: u64, n: u64)

visualize a number of dwords on a given address.

pub fn read_bytes(&mut self, addr: u64, sz: usize) -> PyResult<&[u8]>

read an amount of bytes from an address to a python object.

pub fn read_string_of_bytes(&mut self, addr: u64, sz: usize) -> PyResult<String>

read an amount of bytes from an address to a string of spaced hexa bytes.

pub fn read_string(&self, addr: u64) -> PyResult<String>

read an ascii string from a memory address, if the address point to a non allocated zone string will be empty.

pub fn read_wide_string(&self, addr: u64) -> PyResult<String>

read a wide string from a memory address, if the address point to a non allocated zone string will be empty.

pub fn search_string(&self, kw: &str, map_name: &str) -> PyResult<Vec<u64>>

search a substring on a specific memory map name, it will return a list of matched addresses. if the string is not found, it will return an empty list.

pub fn write_spaced_bytes(
&mut self,
addr: u64,
spaced_hex_bytes: &str
) -> PyResult<bool>

write on emulators memory a spaced hexa bytes

pub fn search_spaced_bytes_from(&self, sbs: &str, saddr: u64) -> PyResult<u64>

search one occurence of a spaced hex bytes from a specific address, will return zero if it’s not found.

pub fn search_spaced_bytes_from_bw(&self, saddr: u64, sbs: &str) -> PyResult<u64>

search one occcurence of a spaced hex bytes from an especific address backward, will return zero if it’s not found.

pub fn search_spaced_bytes(&self, sbs: &str, map_name: &str) -> PyResult<Vec<u64>>

search spaced hex bytes string on specific map using its map name, will return a list with the addresses found if there are matches, otherwise the list will be empty.

pub fn search_spaced_bytes_in_all(&self, sbs: &str) -> PyResult<Vec<u64>>

search spaced hex bytes string on all the memory layout, will return a list with the addresses found if there are matches, otherwise the list will be empty.

pub fn search_string_in_all(&self, kw: String)

Search a substring in all the memory layout except on libs, will print the results. In the future will return a list with results instead of printing.

pub fn search_bytes(&self, bkw: Vec<u8>, map_name: &str) -> PyResult<Vec<u64>>

search a bytes object on specific map, will return a list with matched addresses if there are any.

pub fn allocated_size(&self) -> PyResult<usize>

show the total allocated memory.

pub fn memory_overlaps(&self, addr: u64, sz: u64) -> PyResult<bool>

show if there are memory blocks overlapping eachother.

pub fn show_allocs(&self)

show all the memory blocks allocated during the emulation.

pub fn free(&mut self, name: &str)

free a memory map by its name

pub fn memory_alloc(&self, sz: u64) -> PyResult<u64>

basic allocator, it looks for a free block of given size, it only returns the address if its possible, but dont really allocates, just find the address, you have to load to that address something. use alloc() method instead if possible.

pub fn save_all_allocs(&mut self, path: String)

Save all memory blocks allocated during emulation to disk. Provide a folder where every alloc will be a file.

pub fn save(&mut self, addr: u64, size: u64, filename: String)

save a chunk of memory to disk.

pub fn mem_test(&self) -> PyResult<bool>

perform a memory test to see overlapps or other possible problems.

pub fn bp_show(&self)

breakpoints show breakpoints

pub fn bp_clear_all(&mut self)

clear all the breakpoints

pub fn bp_set_addr(&mut self, addr: u64)

set breakpoint on an address

pub fn bp_get_addr(&self) -> PyResult<u64>

get the current address breakpoint

pub fn bp_set_inst(&mut self, ins: u64)

set breakpoint on a instruction counter

pub fn bp_get_inst(&self) -> PyResult<u64>

get breakpoint on a instrunction counter

pub fn bp_set_mem_read(&mut self, addr: u64)

set a memory breakpoint on read

pub fn bp_get_mem_read(&self) -> PyResult<u64>

get the memory breakpoint on read

pub fn bp_set_mem_write(&mut self, addr: u64)

set a memory breakpoint on write

pub fn bp_get_mem_write(&self) -> PyResult<u64>

get the memory breakpoint on write

pub fn run_hook(&mut self, callback: PyObject)

use a synchronous loop with step() instead of using the run_hook(hook) this method is experimental.